requirements.txt
pendulum==3.1.0
apache-airflow-providers-google

pandas
google-cloud-storage
google-cloud-bigquery

